Backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
bcrypt
requests
python-dotenv
requests
APScheduler